requirements.txt
Flask
Flask-Cors
PyJWT
python-dotenv
werkzeug
pydantic
python-multipart
requests
scikit-learn
numpy
librosa
numba
joblib
gunicorn